+++ src/df-architect.ts
@@ -24,6 +24,7 @@ import WallsCounter from './walls/WallCounter.js';
 import TileCounter from './tiles/TileCounter.js';
 import SoundCounter from './sounds/SoundCounter.js';
 import ShowLayerControls from './general/ShowLayerControls.js';
+import TileConfigExt from './tiles/TileConfigExt.js';
 
 Hooks.once('init', function () {
 	if (!game.modules.get('lib-wrapper')?.active) return;
@@ -53,6 +54,7 @@ Hooks.once('init', function () {
 	try { CaptureGameScreen.init() } catch (error) { console.error(error) }
 	try { LightTemplateManager.init() } catch (error) { console.error(error) }
 	try { TileFlattener.init() } catch (error) { console.error(error) }
+	try { TileConfigExt.init() } catch (error) { console.error(error) }
 });
 
 Hooks.once('setup', function () {